Separate code into tool and util package
